コードレビュー指摘対応
- 非同期実行処理を WarmupService に統合
